requirements.txt
Flask==3.0.0
google-generativeai==0.5.0
stripe==8.0.0
numpy==1.24.0
Werkzeug==3.0.0
gunicorn==21.2.0